# Return uncorrupted EDS from repair, and optimize