Fixes #25.

Optimizes repair logic to do work on a so-far correct EDS struct, and only modify it when modifications are guaranteed to not corrupt.

Note that the original implementation did a quadratic amount of work (deep copy of the entire square) each iteration, for a quadratic number of iterations (i.e. `O(n^4)`). This implementation does linear work (process a single row or column) each iteration, for a quadratic number of iterations (i.e `O(n^3)`). To my understanding, this is order-optimal.

## Initial Results

Bench with `edsBackup` re-added as before #24 against `master` branch as of start of this PR:
```
goos: linux
goarch: amd64
pkg: github.com/lazyledger/rsmt2d
cpu: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
BenchmarkRepair/Repairing_16x16_ODS_using_RSGF8-4         	     261	   4228503 ns/op	 3644160 B/op	   11333 allocs/op
BenchmarkRepair/Repairing_32x32_ODS_using_RSGF8-4         	      61	  19182579 ns/op	16971170 B/op	   42245 allocs/op
BenchmarkRepair/Repairing_64x64_ODS_using_RSGF8-4         	      14	  79352834 ns/op	61468033 B/op	  159749 allocs/op
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-4       	       1	1495244230 ns/op	2933171416 B/op	 1094336 allocs/op
```

Bench without `edsBackup` logic (`master` branch as of the start of this PR):
```
goos: linux
goarch: amd64
pkg: github.com/lazyledger/rsmt2d
cpu: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
BenchmarkRepair/Repairing_16x16_ODS_using_RSGF8-4         	     277	   4127613 ns/op	 3644169 B/op	   11333 allocs/op
BenchmarkRepair/Repairing_32x32_ODS_using_RSGF8-4         	      61	  19092464 ns/op	16971164 B/op	   42245 allocs/op
BenchmarkRepair/Repairing_64x64_ODS_using_RSGF8-4         	      14	  78191005 ns/op	61468040 B/op	  159749 allocs/op
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-4       	       3	 381096647 ns/op	284551544 B/op	  620037 allocs/op
```

We see that removing the deep copy of the EDS at each iteration makes a huge difference at the largest square size. Minor-to-no difference at smaller square sizes. This is mostly within expectations.

Note that the current maximum ODS is 64x64, not 128x128, though that limit is expected to be raised either before mainnet or not too long thereafter.

## Optimized Results

Bench with optimized EDS backup logic:

```
goos: linux
goarch: amd64
pkg: github.com/lazyledger/rsmt2d
cpu: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
BenchmarkRepair/Repairing_16x16_ODS_using_RSGF8-4         	     273	   4292978 ns/op	 3644160 B/op	   11333 allocs/op
BenchmarkRepair/Repairing_32x32_ODS_using_RSGF8-4         	      61	  19140366 ns/op	16971166 B/op	   42245 allocs/op
BenchmarkRepair/Repairing_64x64_ODS_using_RSGF8-4         	      14	  77965243 ns/op	61468058 B/op	  159749 allocs/op
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-4       	       3	 376856007 ns/op	284551544 B/op	  620037 allocs/op
```

We see that the optimized version performs the same as the version without EDS deep copies.