@adlerjohn 2021-04-14
For posterity the leopard tag doesn't work with this branch because of the following bug.

The `Decode` (`Encode`) function for Leopard. The function with `RSGF8` only returns the original shares (parity shares), while the `Decode` (`Encode`) function with `Leopard{FF8,FF16}` returns both the original and parity shares. This was a silent bug because only the first `originalDataWidth` elements of each returned slice were used, so the difference in behavior happened to be ignored. But when doing other manipulation (e.g. `append` the difference in behavior is noticed).

a2a6f15 implements a workaround for half the issue (the decoding half), but can't fully workaround with encoding half.