@liamsi 2021-04-14
> This was a silent bug because only the first originalDataWidth elements of each returned slice were used, so the difference in behavior happened to be ignored. But when doing other manipulation (e.g. append the difference in behavior is noticed).

Does that mean everything works as expected unless you go and modify the returned slices? Couldn't a workaround be to just copy the desired part of the slice before e.g. appending to it? That's what I would do to fix this in go-leopard before diving into the c++ code of the actual leopard library again.