@adlerjohn 2021-04-14
@liamsi did a full workaround, so things work now with Leopard.